#161 Moved more code out of gtfs_stats
