+++ gtfs/gtfs_utils/gtfs_utils/gtfs_stats.py
@@ -6,11 +6,7 @@
 # MoT GTFS archive and schedule for nightly runs. 
 
 import pandas as pd
-import numpy as np
 import datetime
-import gtfs_utils as gu
-import gtfstk
-from collections import OrderedDict, defaultdict
 import os
 from os.path import join
 import boto3
@@ -20,457 +16,10 @@ from tqdm import tqdm
 from partridge import feed as ptg_feed
 from botocore.handlers import disable_signing
 from gtfs_stats_conf import *
-from retry import retry
 from environment import init_conf
 from general_utils import parse_date
-
-
-def compute_trip_stats_partridge(feed, zones):
-    """
-    Parameters
-    ----------
-    feed : partridge feed
-    zones: DataFrame with stop_code to zone_name mapping
-    
-    Returns
-    -------
-    DataFrame with the following columns:
-
-    - ``'trip_id'``
-    - ``'route_id'``
-    - ``'route_short_name'``
-    - ``'route_short_name'``
-    - ``'agency_id'``
-    - ``'agency_name'``
-    - ``'route_long_name'``
-    - ``'route_type'``
-    - ``'direction_id'``
-    - ``'shape_id'``
-    - ``'num_stops'``: number of stops on trip
-    - ``'start_time'``: first departure time of the trip
-    - ``'end_time'``: last departure time of the trip
-    - ``'start_stop_id'``: stop ID of the first stop of the trip
-    - ``'end_stop_id'``: stop ID of the last stop of the trip
-    - ``'start_stop_name'``: stop name of the first stop of the trip
-    - ``'end_stop_name'``: stop name of the last stop of the trip
-    - ``'start_stop_code'``: stop code of the first stop of the trip
-    - ``'end_stop_code'``: stop code of the last stop of the trip
-    - ``'start_stop_lat'``: ``start_stop_lat`` of the first stop of the trip
-    - ``'start_stop_lon'``: ``start_stop_lon`` of the first stop of the trip
-    - ``'end_stop_lat'``: ``end_stop_lat`` of the last stop of the trip
-    - ``'end_stop_lon'``: ``end_stop_lon`` of the last stop of the trip
-    - ``'start_zone'``: zone name of the first stop of the trip
-    - ``'end_zone'``: zone name of the last stop of the trip
-    - ``'num_zones'``:  ``num_zones`` of the first stop of the trip
-    - ``'num_zones_missing'``:  ``num_zones_missing`` of the first stop of the trip
-    - ``'is_loop'``: 1 if the start and end stop are less than 400m apart and
-      0 otherwise
-    - ``'distance'``: distance of the trip in ``feed.dist_units``;
-      contains all ``np.nan`` entries if ``feed.shapes is None``
-    - ``'duration'``: duration of the trip in hours
-    - ``'speed'``: distance/duration
-
-    TODO: this is not true here, we're only using shape_dist_traveled
-    TODO: implement or drop from docs
-    If ``feed.stop_times`` has a ``shape_dist_traveled`` column with at
-    least one non-NaN value and ``compute_dist_from_shapes == False``,
-    then use that column to compute the distance column.
-    Else if ``feed.shapes is not None``, then compute the distance
-    column using the shapes and Shapely.
-    Otherwise, set the distances to NaN.
-
-    If route IDs are given, then restrict to trips on those routes.
-
-    Notes
-    -----
-    - Assume the following feed attributes are not ``None``:
-
-        * ``feed.trips``
-        * ``feed.routes``
-        * ``feed.stop_times``
-        * ``feed.shapes`` (optionally)
-        * Those used in :func:`.stops.build_geometry_by_stop`
-
-    - Calculating trip distances with ``compute_dist_from_shapes=True``
-      seems pretty accurate.  For example, calculating trip distances on
-      `this Portland feed
-      <https://transitfeeds.com/p/trimet/43/1400947517>`_
-      using ``compute_dist_from_shapes=False`` and
-      ``compute_dist_from_shapes=True``,
-      yields a difference of at most 0.83km from the original values.
-
-    """
-    f = feed.trips
-    f = (f[['route_id', 'trip_id', 'direction_id', 'shape_id']]
-         .merge(feed.routes[['route_id', 'route_short_name', 'route_long_name',
-                             'route_type', 'agency_id', 'route_desc']])
-         .merge(feed.agency[['agency_id', 'agency_name']], how='left', on='agency_id')
-         .merge(feed.stop_times)
-         .merge(feed.stops[['stop_id', 'stop_name', 'stop_lat', 'stop_lon', 'stop_code', 'stop_desc']])
-         .merge(zones, how='left')
-         .sort_values(['trip_id', 'stop_sequence'])
-         # .assign(departure_time=lambda x: x['departure_time'].map(
-         #    hp.timestr_to_seconds)
-         #       )
-         )
-
-    # parse route_desc
-    f[['route_mkt', 'route_direction', 'route_alternative']] = f['route_desc'].str.split('-', expand=True)
-    f = f.drop('route_desc', axis=1)
-
-    # parse stop_desc
-    stop_desc_fields = {'street': 'רחוב',
-                        'city': 'עיר',
-                        'platform': 'רציף',
-                        'floor': 'קומה'}
-
-    stop_desc_prefix = 'stop_desc_'
-
-    STOP_DESC_RE = ''
-    for n, fld in stop_desc_fields.items():
-        STOP_DESC_RE += fld + f':(?P<{stop_desc_prefix + n}>.*)'
-
-    sd = f.stop_desc.str.extract(STOP_DESC_RE).apply(lambda x: x.str.strip())
-    f = pd.concat([f, sd], axis=1)
-
-    # get geometry by stop for distance measurement
-    geometry_by_stop = gtfstk.build_geometry_by_stop(feed, use_utm=True)
-
-    g = f.groupby('trip_id')
-
-    def my_agg(group):
-        d = OrderedDict()
-        d['route_id'] = group['route_id'].iat[0]
-        d['route_short_name'] = group['route_short_name'].iat[0]
-        d['route_long_name'] = group['route_long_name'].iat[0]
-        d['route_mkt'] = group['route_mkt'].iat[0]
-        d['route_direction'] = group['route_direction'].iat[0]
-        d['route_alternative'] = group['route_alternative'].iat[0]
-        d['agency_id'] = group['agency_id'].iat[0]
-        d['agency_name'] = group['agency_name'].iat[0]
-        d['route_type'] = group['route_type'].iat[0]
-        d['direction_id'] = group['direction_id'].iat[0]
-        d['shape_id'] = group['shape_id'].iat[0]
-        d['num_stops'] = group.shape[0]
-        d['start_time'] = group['departure_time'].iat[0]
-        d['end_time'] = group['departure_time'].iat[-1]
-        d['start_stop_id'] = group['stop_id'].iat[0]
-        d['end_stop_id'] = group['stop_id'].iat[-1]
-        d['start_stop_code'] = group['stop_code'].iat[0]
-        d['end_stop_code'] = group['stop_code'].iat[-1]
-        d['start_stop_name'] = group['stop_name'].iat[0]
-        d['end_stop_name'] = group['stop_name'].iat[-1]
-        d['start_stop_desc'] = group['stop_desc'].iat[0]
-        d['end_stop_desc'] = group['stop_desc'].iat[-1]
-        d['start_stop_lat'] = group['stop_lat'].iat[0]
-        d['start_stop_lon'] = group['stop_lon'].iat[0]
-        d['end_stop_lat'] = group['stop_lat'].iat[-1]
-        d['end_stop_lon'] = group['stop_lon'].iat[-1]
-        d['start_stop_city'] = group['stop_desc_city'].iat[0]
-        d['end_stop_city'] = group['stop_desc_city'].iat[-1]
-        d['start_zone'] = group['zone_name'].iat[0]
-        d['end_zone'] = group['zone_name'].iat[-1]
-        d['num_zones'] = group.zone_name.nunique()
-        d['num_zones_missing'] = group.zone_name.isnull().sum()
-        dist = geometry_by_stop[d['start_stop_id']].distance(
-            geometry_by_stop[d['end_stop_id']])
-        d['is_loop'] = int(dist < 400)
-        d['duration'] = (d['end_time'] - d['start_time']) / 3600
-
-        d['all_stop_latlon'] = ';'.join(str(x)+','+str(y) for x, y in
-                                        zip(group['stop_lat'].tolist(), group['stop_lon'].tolist()))
-
-        d['all_stop_code'] = ';'.join(group['stop_code'].tolist())
-        d['all_stop_id'] = ';'.join(group['stop_id'].tolist())
-        d['all_stop_desc_city'] = ';'.join(group['stop_desc_city'].tolist())
-
-        return pd.Series(d)
-
-    h = g.apply(my_agg)
-    h['distance'] = g.shape_dist_traveled.max()
-
-    # Reset index and compute final stats
-    h = h.reset_index()
-    h['speed'] = h['distance'] / h['duration'] / 1000
-    h[['start_time', 'end_time']] = (
-        h[['start_time', 'end_time']].applymap(
-            lambda x: gtfstk.helpers.timestr_to_seconds(x, inverse=True))
-    )
-    return h
-
-
-def get_active_trips_df(trip_times):
-    """
-    Count the number of trips in ``trip_times`` that are active
-    at any given time.
-
-    Parameters
-    ----------
-    trip_times : DataFrame
-        Contains columns
-
-        - start_time: start time of the trip in seconds past midnight
-        - end_time: end time of the trip in seconds past midnight
-
-    Returns
-    -------
-    Series
-        index is times from midnight when trips start and end, values are number of active trips for that time
-
-    """
-    active_trips = pd.concat([pd.Series(1, trip_times.start_time),  # departed add 1
-                              pd.Series(-1, trip_times.end_time)  # arrived subtract 1
-                              ]).groupby(level=0, sort=True).sum().cumsum().ffill()
-    return active_trips
-
-
-def compute_route_stats_base_partridge(trip_stats_subset,
-                                       headway_start_time='07:00:00',
-                                       headway_end_time='19:00:00',
-                                       *,
-                                       split_directions=False):
-    """
-    Compute stats for the given subset of trips stats.
-
-    Parameters
-    ----------
-    trip_stats_subset : DataFrame
-        Subset of the output of :func:`.trips.compute_trip_stats`
-    split_directions : boolean
-        If ``True``, then separate the stats by trip direction (0 or 1);
-        otherwise aggregate trips visiting from both directions. 
-        Default: ``False``
-    headway_start_time : string
-        HH:MM:SS time string indicating the start time for computing
-        headway stats
-        Default: ``'07:00:00'``
-    headway_end_time : string
-        HH:MM:SS time string indicating the end time for computing
-        headway stats.
-        Default: ``'19:00:00'``
-
-    Returns
-    -------
-    DataFrame
-        Columns are
-
-        - ``'route_id'``
-        - ``'route_short_name'``
-        - ``'agency_id'``
-        - ``'agency_name'``
-        - ``'route_long_name'``
-        - ``'route_type'``
-        - ``'direction_id'``: 1/0
-        - ``'num_trips'``: number of trips on the route in the subset
-        - ``'num_trip_starts'``: number of trips on the route with
-          nonnull start times
-        - ``'num_trip_ends'``: number of trips on the route with nonnull
-          end times that end before 23:59:59
-        - ``'is_loop'``: 1 if at least one of the trips on the route has
-          its ``is_loop`` field equal to 1; 0 otherwise
-        - ``'is_bidirectional'``: 1 if the route has trips in both
-          directions; 0 otherwise
-        - ``'start_time'``: start time of the earliest trip on the route
-        - ``'end_time'``: end time of latest trip on the route
-        - ``'max_headway'``: maximum of the durations (in minutes)
-          between trip starts on the route between
-          ``headway_start_time`` and ``headway_end_time`` on the given
-          dates
-        - ``'min_headway'``: minimum of the durations (in minutes)
-          mentioned above
-        - ``'mean_headway'``: mean of the durations (in minutes)
-          mentioned above
-        - ``'peak_num_trips'``: maximum number of simultaneous trips in
-          service (for the given direction, or for both directions when
-          ``split_directions==False``)
-        - ``'peak_start_time'``: start time of first longest period
-          during which the peak number of trips occurs
-        - ``'peak_end_time'``: end time of first longest period during
-          which the peak number of trips occurs
-        - ``'service_duration'``: total of the duration of each trip on
-          the route in the given subset of trips; measured in hours
-        - ``'service_distance'``: total of the distance traveled by each
-          trip on the route in the given subset of trips; measured in
-          whatever distance units are present in ``trip_stats_subset``;
-          contains all ``np.nan`` entries if ``feed.shapes is None``
-        - ``'service_speed'``: service_distance/service_duration;
-          measured in distance units per hour
-        - ``'mean_trip_distance'``: service_distance/num_trips
-        - ``'mean_trip_duration'``: service_duration/num_trips
-        - ``'start_stop_id'``: ``start_stop_id`` of the first trip for the route
-        - ``'end_stop_id'``: ``end_stop_id`` of the first trip for the route
-        - ``'start_stop_lat'``: ``start_stop_lat`` of the first trip for the route
-        - ``'start_stop_lon'``: ``start_stop_lon`` of the first trip for the route
-        - ``'end_stop_lat'``: ``end_stop_lat`` of the first trip for the route
-        - ``'end_stop_lon'``: ``end_stop_lon`` of the first trip for the route
-        - ``'num_stops'``: ``num_stops`` of the first trip for the route
-        - ``'start_zone'``: ``start_zone`` of the first trip for the route
-        - ``'end_zone'``: ``end_zone`` of the first trip for the route
-        - ``'num_zones'``:  ``num_zones`` of the first trip for the route
-        - ``'num_zones_missing'``:  ``num_zones_missing`` of the first trip for the route
-
-        TODO: actually implement split_directions
-        If not ``split_directions``, then remove the
-        direction_id column and compute each route's stats,
-        except for headways, using
-        its trips running in both directions.
-        In this case, (1) compute max headway by taking the max of the
-        max headways in both directions; (2) compute mean headway by
-        taking the weighted mean of the mean headways in both
-        directions.
-        
-        If ``trip_stats_subset`` is empty, return an empty DataFrame.
-
-    """
-    f = trip_stats_subset.copy()
-    f[['start_time', 'end_time']] = f[['start_time', 'end_time']].applymap(gtfstk.helpers.timestr_to_seconds)
-
-    headway_start = gtfstk.helpers.timestr_to_seconds(headway_start_time)
-    headway_end = gtfstk.helpers.timestr_to_seconds(headway_end_time)
-
-    def compute_route_stats(group):
-        d = OrderedDict()
-        d['route_short_name'] = group['route_short_name'].iat[0]
-        d['route_long_name'] = group['route_long_name'].iat[0]
-        d['route_mkt'] = group['route_mkt'].iat[0]
-        d['route_direction'] = group['route_direction'].iat[0]
-        d['route_alternative'] = group['route_alternative'].iat[0]
-        d['agency_id'] = group['agency_id'].iat[0]
-        d['agency_name'] = group['agency_name'].iat[0]
-        d['route_type'] = group['route_type'].iat[0]
-        d['num_trips'] = group.shape[0]
-        d['num_trip_starts'] = group['start_time'].count()
-        d['num_trip_ends'] = group.loc[
-            group['end_time'] < 24 * 3600, 'end_time'].count()
-        d['is_loop'] = int(group['is_loop'].any())
-        d['is_bidirectional'] = int(group['direction_id'].unique().size > 1)
-        d['start_time'] = group['start_time'].min()
-        d['end_time'] = group['end_time'].max()
-
-        # Compute headway stats
-        headways = np.array([])
-        for direction in [0, 1]:
-            stimes = group[group['direction_id'] == direction][
-                'start_time'].values
-            stimes = sorted([stime for stime in stimes
-                             if headway_start <= stime <= headway_end])
-            headways = np.concatenate([headways, np.diff(stimes)])
-        if headways.size:
-            d['max_headway'] = np.max(headways) / 60  # minutes
-            d['min_headway'] = np.min(headways) / 60  # minutes
-            d['mean_headway'] = np.mean(headways) / 60  # minutes
-        else:
-            d['max_headway'] = np.nan
-            d['min_headway'] = np.nan
-            d['mean_headway'] = np.nan
-
-        # Compute peak num trips
-        active_trips = get_active_trips_df(group[['start_time', 'end_time']])
-        times, counts = active_trips.index.values, active_trips.values
-        start, end = gtfstk.helpers.get_peak_indices(times, counts)
-        d['peak_num_trips'] = counts[start]
-        d['peak_start_time'] = times[start]
-        d['peak_end_time'] = times[end]
-
-        d['service_distance'] = group['distance'].sum()
-        d['service_duration'] = group['duration'].sum()
-
-        # Added by cjer
-        d['start_stop_id'] = group['start_stop_id'].iat[0]
-        d['end_stop_id'] = group['end_stop_id'].iat[0]
-        d['start_stop_name'] = group['start_stop_name'].iat[0]
-        d['end_stop_name'] = group['end_stop_name'].iat[0]
-        d['start_stop_desc'] = group['start_stop_desc'].iat[0]
-        d['end_stop_desc'] = group['end_stop_desc'].iat[0]
-
-        d['start_stop_lat'] = group['start_stop_lat'].iat[0]
-        d['start_stop_lon'] = group['start_stop_lon'].iat[0]
-        d['end_stop_lat'] = group['end_stop_lat'].iat[0]
-        d['end_stop_lon'] = group['end_stop_lon'].iat[0]
-
-        d['start_stop_city'] = group['start_stop_city'].iat[0]
-        d['end_stop_city'] = group['end_stop_city'].iat[0]
-
-        d['num_stops'] = group['num_stops'].iat[0]
-
-        d['start_zone'] = group['start_zone'].iat[0]
-        d['end_zone'] = group['end_zone'].iat[0]
-        d['num_zones'] = group['num_zones'].iat[0]
-        d['num_zones_missing'] = group['num_zones_missing'].iat[0]
-        d['all_stop_latlon'] = group['all_stop_latlon'].iat[0]
-
-        d['all_stop_code'] = group['all_stop_code'].iat[0]
-        d['all_stop_id'] = group['all_stop_id'].iat[0]
-        d['all_stop_desc_city'] = group['all_stop_desc_city'].iat[0]
-
-        d['all_start_time'] = ';'.join([gtfstk.helpers.timestr_to_seconds(x, inverse=True)
-                                       for x in group['start_time'].tolist()])
-        d['all_trip_id'] = ';'.join(group['trip_id'].tolist())
-        return pd.Series(d)
-
-    g = f.groupby('route_id').apply(
-        compute_route_stats).reset_index()
-
-    # Compute a few more stats
-    g['service_speed'] = g['service_distance'] / g['service_duration']
-    g['mean_trip_distance'] = g['service_distance'] / g['num_trips']
-    g['mean_trip_duration'] = g['service_duration'] / g['num_trips']
-
-    # Convert route times to time strings
-    time_cols = ['start_time', 'end_time', 'peak_start_time', 'peak_end_time']
-    g[time_cols] = g[time_cols].applymap(lambda x: gtfstk.helpers.timestr_to_seconds(x, inverse=True))
-
-    g['service_speed'] = g.service_speed / 1000
-
-    g = g[['route_id', 'route_short_name', 'agency_id', 'agency_name',
-           'route_long_name', 'route_type', 'route_mkt', 'route_direction', 'route_alternative',
-           'num_trips', 'num_trip_starts',
-           'num_trip_ends', 'is_loop', 'is_bidirectional', 'start_time',
-           'end_time', 'max_headway', 'min_headway', 'mean_headway',
-           'peak_num_trips', 'peak_start_time', 'peak_end_time',
-           'service_distance', 'service_duration', 'service_speed',
-           'mean_trip_distance', 'mean_trip_duration', 'start_stop_id',
-           'end_stop_id', 'start_stop_name', 'end_stop_name',
-           'start_stop_city', 'end_stop_city',
-           'start_stop_desc', 'end_stop_desc',
-           'start_stop_lat', 'start_stop_lon', 'end_stop_lat',
-           'end_stop_lon', 'num_stops', 'start_zone', 'end_zone',
-           'num_zones', 'num_zones_missing',
-           'all_stop_latlon', 'all_stop_code', 'all_stop_id', 'all_stop_desc_city', 'all_start_time', 'all_trip_id',
-           ]]
-
-    return g
-
-@retry()
-def s3_download(bucket, key, output_path):
-    """
-Download file from s3 bucket. Retry using decorator, and report to logger given in report parameter.
-    :param bucket: s3 boto bucket object
-    :type bucket: boto3.resources.factory.s3.Bucket
-    :param key: key of the file to download
-    :type key: str
-    :param output_path: output path to download the file to
-    :type output_path: str
-    """
-
-    def hook(t):
-        def inner(bytes_amount):
-            t.update(bytes_amount)
-
-        return inner
-
-    if DOWNLOAD_PBAR:
-        # TODO: this is an S3 anti-pattern, and is inefficient - so defaulting to not doing this
-        if SIZE_FOR_DOWNLOAD_PBAR:
-            size = [obj.size for obj in bucket.objects.filter(Prefix=key, MaxKeys=1)][0]
-        else:
-            size = None
-
-        with tqdm(total=size, unit='B', unit_scale=True, desc='download ' + key, leave=True) as t:
-            bucket.download_file(key, output_path, Callback=hook(t))
-    else:
-        bucket.download_file(key, output_path)
-
+from s3 import get_valid_file_dates_dict, s3_download
+import gtfs_utils as gu
 
 def _get_existing_output_files(output_folder):
     """
@@ -483,82 +32,6 @@ Get existing output files in the given folder, in a list containing tuples of da
     return [(g[0], g[1]) for g in
             (re.match(OUTPUT_FILE_NAME_RE, file).groups()
              for file in os.listdir(output_folder))]
-
-
-def get_bucket_valid_files(bucket_objects):
-    """
-Get list of valid files from bucket, as set by BUCKET_VALID_FILES_RE
-    :param bucket_objects: collection of bucket objects
-    :type bucket_objects: s3.Bucket.objectsCollection
-    :return: list of valid file keys
-    :rtype: list of str
-    """
-    return [obj.key for obj in bucket_objects
-            if re.match(BUCKET_VALID_FILES_RE, obj.key)]
-
-
-def get_dates_without_output(valid_files, existing_output_files):
-    """
-Get list of dates without output files (currently just route_stats is considered)
-    :param valid_files: list of valid file keys
-    :rtype: list of str
-    :param existing_output_files: list of 2-tuples as returned by _get_existing_output_files
-    :type existing_output_files: list
-    :return: list of valid file keys for stat computation
-    :rtype: list
-    """
-    return [parse_date(file)[1] for file in valid_files
-            if file not in [g[0] + '.zip'
-                            for g in existing_output_files
-                            if g[1] == 'route_stats']]
-
-
-def get_forward_fill_dict(valid_files, future_days=FUTURE_DAYS):
-    """
-get a dictionary mapping gtfs file names to a list of dates for forward fill by scanning for missing dates for files
-    :param valid_files: list of valid file keys
-    :rtype: list of str
-    :return dictionary mapping file names to dates
-    :rtype defaultdict of lists (defaults to empty list)
-    """
-    existing_dates = pd.DatetimeIndex([parse_date(file)[0] for file in valid_files])
-    expected_dates = pd.DatetimeIndex(start=existing_dates.min(), end=existing_dates.max()+datetime.timedelta(days=future_days), freq='D')
-    date_df = pd.Series(pd.NaT, expected_dates)
-    date_df[existing_dates] = existing_dates
-    date_df = date_df.fillna(method='ffill', limit=59)
-    # TODO: remove dates that aren't in the 59 day gap
-    # BUG: will act unexcpectedly if more than 59 day gap
-    ffill = defaultdict(list)
-    for file_date, stats_date in zip(date_df.dt.strftime('%Y-%m-%d'), date_df.index.strftime('%Y-%m-%d')):
-        ffill[file_date + '.zip'].append(stats_date)
-
-    return ffill
-
-
-def get_valid_file_dates_dict(bucket_objects, existing_output_files, logger,
-                              forward_fill):
-    logger.info(f'BUCKET_VALID_FILES_RE={BUCKET_VALID_FILES_RE}')
-    bucket_valid_files = get_bucket_valid_files(bucket_objects)
-    if forward_fill:
-        logger.info(f'applying forward fill')
-        ffill_dict = get_forward_fill_dict(bucket_valid_files)
-        logger.info(f'found {sum([len(l) for l in ffill_dict.values()]) - len(bucket_valid_files)} missing dates for '
-                    'forward fill.')
-
-        files_for_stats = defaultdict(list)
-        for file in ffill_dict:
-            files_for_stats[file] = get_dates_without_output([date_str + '.zip' for date_str in ffill_dict[file]],
-                                                             existing_output_files)
-
-    else:
-        files_for_stats = defaultdict(list)
-        for date in get_dates_without_output(bucket_valid_files, existing_output_files):
-            files_for_stats[date + '.zip'].append(date)
-
-    logger.info(f'found {len([key for key in files_for_stats if len(files_for_stats[key])>0])} GTFS files valid for '
-                'stats calculations in bucket')
-    logger.debug(f'Files: { {key: value for key, value in files_for_stats.items() if len(files_for_stats[key])>0} }')
-    return files_for_stats
 
 
 def get_gtfs_file(file, gtfs_folder, bucket, logger, force=False):
@@ -654,7 +127,7 @@ and route_stats).
         logger.info(
             f'starting compute_trip_stats_partridge for file "{join(gtfs_folder, file)}" with date "{date}" and zones '
             f'"{LOCAL_TARIFF_PATH}"')
-        ts = compute_trip_stats_partridge(feed, zones)
+        ts = gu.compute_trip_stats_partridge(feed, zones)
         logger.debug(
             f'finished compute_trip_stats_partridge for file "{join(gtfs_folder, file)}" with date "{date}" and zones '
             f'"{LOCAL_TARIFF_PATH}"')
@@ -671,7 +144,7 @@ and route_stats).
         f'num_start_zones={ts.start_zone.nunique()}, num_agency={ts.agency_name.nunique()}')
 
     logger.info(f'starting compute_route_stats_base_partridge from trip stats result')
-    rs = compute_route_stats_base_partridge(ts)
+    rs = gu.compute_route_stats_base_partridge(ts)
     logger.debug(f'finished compute_route_stats_base_partridge from trip stats result')
     # TODO: log this
     rs['date'] = date_str
@@ -761,12 +234,10 @@ Will look for downloaded GTFS feeds with matching names in given gtfs_folder.
         bucket_objects = bucket.objects.all()
         logger.info(f'number of objects in bucket: {sum(1 for _ in bucket_objects)}')
 
-        if forward_fill:
-            file_dates_dict = get_valid_file_dates_dict(bucket_objects, existing_output_files, logger,
-                                                        forward_fill=True)
-        else:
-            file_dates_dict = get_valid_file_dates_dict(bucket_objects, existing_output_files, logger,
-                                                        forward_fill=False)
+        file_dates_dict = get_valid_file_dates_dict(bucket_objects,
+                                                    existing_output_files,
+                                                    logger,
+                                                    forward_fill=forward_fill)
 
         non_empty_file_dates = {key: value for key, value in file_dates_dict.items() if len(file_dates_dict[key]) > 0}
         with tqdm(non_empty_file_dates, postfix='initializing', unit='file', desc='files') as t:
